requirements-dev.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --extra=dev --output-file=requirements-dev.txt
#
asyncache==0.3.1
    # via iceprod
attrs==22.2.0
    # via pytest
backoff==2.2.1
    # via opentelemetry-exporter-otlp-proto-http
bitmath==1.3.3.1
    # via wipac-file-catalog-indexer (setup.py)
boto3==1.26.95
    # via iceprod
botocore==1.29.95
    # via
    #   boto3
    #   s3transfer
cachetools==5.3.0
    # via
    #   asyncache
    #   iceprod
    #   wipac-rest-tools
certifi==2022.12.7
    # via
    #   iceprod
    #   requests
cffi==1.15.1
    # via cryptography
charset-normalizer==3.1.0
    # via requests
click==8.1.3
    # via
    #   click-completion
    #   pycycle
click-completion==0.5.2
    # via pycycle
colorama==0.4.6
    # via crayons
coloredlogs==15.0.1
    # via
    #   wipac-file-catalog
    #   wipac-file-catalog-indexer (setup.py)
    #   wipac-telemetry
coverage[toml]==7.2.2
    # via pytest-cov
crayons==0.4.0
    # via pycycle
cryptography==39.0.2
    # via
    #   iceprod
    #   pyjwt
    #   pyopenssl
deprecated==1.2.13
    # via opentelemetry-api
dnspython==2.3.0
    # via pymongo
exceptiongroup==1.1.1
    # via pytest
flake8==6.0.0
    # via wipac-file-catalog-indexer (setup.py)
googleapis-common-protos==1.56.2
    # via
    #   opentelemetry-exporter-jaeger-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
grpcio==1.51.3
    # via opentelemetry-exporter-jaeger-proto-grpc
humanfriendly==10.0
    # via coloredlogs
iceprod==2.6.8
    # via wipac-file-catalog-indexer (setup.py)
idna==3.4
    # via requests
iniconfig==2.0.0
    # via pytest
jinja2==3.1.2
    # via click-completion
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
ldap3==2.9.1
    # via
    #   iceprod
    #   wipac-file-catalog
markupsafe==2.1.2
    # via jinja2
mccabe==0.7.0
    # via flake8
more-itertools==9.1.0
    # via wipac-file-catalog-indexer (setup.py)
motor==3.1.1
    # via
    #   iceprod
    #   wipac-file-catalog
mypy==1.1.1
    # via wipac-file-catalog-indexer (setup.py)
mypy-extensions==1.0.0
    # via mypy
natsort==8.3.1
    # via wipac-file-catalog-indexer (setup.py)
opentelemetry-api==1.16.0
    # via
    #   opentelemetry-exporter-jaeger-proto-grpc
    #   opentelemetry-exporter-jaeger-thrift
    #   opentelemetry-exporter-otlp-proto-http
    #   opentelemetry-sdk
    #   wipac-telemetry
opentelemetry-exporter-jaeger==1.16.0
    # via wipac-telemetry
opentelemetry-exporter-jaeger-proto-grpc==1.16.0
    # via opentelemetry-exporter-jaeger
opentelemetry-exporter-jaeger-thrift==1.16.0
    # via opentelemetry-exporter-jaeger
opentelemetry-exporter-otlp-proto-http==1.16.0
    # via wipac-telemetry
opentelemetry-proto==1.16.0
    # via opentelemetry-exporter-otlp-proto-http
opentelemetry-sdk==1.16.0
    # via
    #   opentelemetry-exporter-jaeger-proto-grpc
    #   opentelemetry-exporter-jaeger-thrift
    #   opentelemetry-exporter-otlp-proto-http
    #   wipac-telemetry
opentelemetry-semantic-conventions==0.37b0
    # via opentelemetry-sdk
packaging==23.0
    # via pytest
pluggy==1.0.0
    # via pytest
protobuf==3.20.3
    # via
    #   googleapis-common-protos
    #   opentelemetry-proto
    #   wipac-telemetry
psutil==5.9.4
    # via iceprod
pyasn1==0.4.8
    # via ldap3
pycodestyle==2.10.0
    # via flake8
pycparser==2.21
    # via cffi
pycycle==0.0.8
    # via wipac-file-catalog-indexer (setup.py)
pyflakes==3.0.1
    # via flake8
pyjwt[crypto]==2.5.0
    # via wipac-rest-tools
pymongo==4.3.3
    # via
    #   iceprod
    #   motor
    #   wipac-file-catalog
pymysql==1.0.2
    # via wipac-file-catalog-indexer (setup.py)
pyopenssl==23.0.0
    # via iceprod
pypng==0.20220715.0
    # via qrcode
pytest==7.2.2
    # via
    #   pycycle
    #   pytest-asyncio
    #   pytest-cov
    #   pytest-mock
    #   wipac-file-catalog-indexer (setup.py)
pytest-asyncio==0.21.0
    # via wipac-file-catalog-indexer (setup.py)
pytest-cov==4.0.0
    # via wipac-file-catalog-indexer (setup.py)
pytest-mock==3.10.0
    # via wipac-file-catalog-indexer (setup.py)
python-dateutil==2.8.2
    # via
    #   botocore
    #   iceprod
pyyaml==6.0
    # via iceprod
qrcode==7.4.2
    # via wipac-rest-tools
requests==2.28.2
    # via
    #   iceprod
    #   opentelemetry-exporter-otlp-proto-http
    #   requests-futures
    #   requests-mock
    #   requests-toolbelt
    #   wipac-dev-tools
    #   wipac-file-catalog
    #   wipac-file-catalog-indexer (setup.py)
    #   wipac-rest-tools
requests-futures==1.0.0
    # via
    #   iceprod
    #   wipac-file-catalog
    #   wipac-file-catalog-indexer (setup.py)
    #   wipac-rest-tools
requests-mock==1.10.0
    # via wipac-file-catalog-indexer (setup.py)
requests-toolbelt==0.10.1
    # via
    #   iceprod
    #   wipac-file-catalog
    #   wipac-file-catalog-indexer (setup.py)
ruff==0.0.257
    # via wipac-file-catalog-indexer (setup.py)
s3transfer==0.6.0
    # via boto3
setproctitle==1.3.2
    # via iceprod
shellingham==1.5.0.post1
    # via click-completion
six==1.16.0
    # via
    #   click-completion
    #   python-dateutil
    #   requests-mock
    #   thrift
statsd==4.0.1
    # via iceprod
thrift==0.16.0
    # via opentelemetry-exporter-jaeger-thrift
tomli==2.0.1
    # via
    #   coverage
    #   mypy
    #   pytest
tornado==6.2
    # via
    #   iceprod
    #   wipac-file-catalog
    #   wipac-rest-tools
types-cryptography==3.3.23.2
    # via pyjwt
types-pymysql==1.0.19.5
    # via wipac-file-catalog-indexer (setup.py)
types-pyyaml==6.0.12.8
    # via wipac-file-catalog-indexer (setup.py)
types-requests==2.28.11.15
    # via wipac-file-catalog-indexer (setup.py)
types-urllib3==1.26.25.8
    # via types-requests
typing-extensions==4.5.0
    # via
    #   mypy
    #   opentelemetry-sdk
    #   qrcode
    #   wipac-dev-tools
    #   wipac-file-catalog-indexer (setup.py)
    #   wipac-telemetry
unidecode==1.3.6
    # via iceprod
urllib3==1.26.15
    # via
    #   botocore
    #   requests
wipac-dev-tools==1.6.13
    # via
    #   iceprod
    #   wipac-rest-tools
    #   wipac-telemetry
wipac-file-catalog==1.9.3
    # via wipac-file-catalog-indexer (setup.py)
wipac-rest-tools[telemetry]==1.4.16
    # via
    #   iceprod
    #   wipac-file-catalog
    #   wipac-file-catalog-indexer (setup.py)
wipac-telemetry==0.2.5
    # via wipac-rest-tools
wrapt==1.15.0
    # via deprecated
xmltodict==0.13.0
    # via wipac-file-catalog-indexer (setup.py)

# The following packages are considered to be unsafe in a requirements file:
# setuptools
